+++ main.py
@@ -10,20 +10,20 @@ def main():
     juego = Juego(jugador1, jugador2)
     juego.empezar_juego()
     
-    while juego.estado() == True:
+    while juego.estado == True:
         
-        juego.tablero()
+        juego.tablero.imprimir_tablero()
+        
         print(f'Ahora es el turno de {juego.__turno__}')
-        
         pieza = input('Que pieza quieres mover?')
+        
         if pieza == '0':
             juego.terminar_juego()
         else:    
             fila = int(input('Fila donde quieres mover la pieza: '))
             columna = int(input('Columna donde quieres mover la pieza: '))
-            juego.__tablero__.mover_pieza(fila, columna, pieza)
-
-            juego.__tablero__.imprimir_tablero()
+            juego.tablero.mover_pieza(fila, columna, pieza)
+            juego.tablero.imprimir_tablero()
             juego.cambiar_turno()